requirements.txt
numpy==1.19.5
pandas==1.1.5
seaborn==0.11.2
scikit-learn==0.24.2
torch==1.9.0
transformers==4.9.2
pytorch-ignite==0.4.7